Fix linting and unit tests
