+++ backend/src/graphql/models/InvitedTable.ts
@@ -1,0 +1,13 @@
+import { Field, ObjectType } from 'type-graphql'
+
+import { OpenTable } from './TableModel'
+import { User } from './UserModel'
+
+@ObjectType()
+export class InvitedTable {
+  @Field(() => OpenTable)
+  table: OpenTable
+
+  @Field(() => User)
+  user: User
+}
