+++ backend/src/graphql/pubSub.ts
@@ -1,7 +1,7 @@
 import { createPubSub } from 'graphql-yoga'
 
 import { MeetingInfo } from '#api/BBB'
-import { InvitedTable } from '#models/TableModel'
+import { InvitedTable } from '#models/InvitedTable'
 
 export const pubSub = createPubSub<{
   OPEN_TABLE_SUBSCRIPTION: [MeetingInfo[]]
