+++ frontend/src/stores/tablesStore.spec.ts
@@ -4,14 +4,17 @@ import { setActivePinia, createPinia } from 'pinia'
 import { vi, describe, it, expect, beforeEach } from 'vitest'
 
 import { tablesQuery } from '#queries/tablesQuery'
+import { inviteTableSubscription } from '#subscriptions/inviteTableSubscription'
 import { updateOpenTablesSubscription } from '#subscriptions/updateOpenTablesSubscription'
 
 import { useTablesStore } from './tablesStore'
 
 const updateTablesSubscriptionMock: IMockSubscription = createMockSubscription()
+const inviteTableSubscriptionMock: IMockSubscription = createMockSubscription()
 const mockClient = createMockClient()
 const tablesQueryMock = vi.fn()
 mockClient.setRequestHandler(updateOpenTablesSubscription, () => updateTablesSubscriptionMock)
+mockClient.setRequestHandler(inviteTableSubscription, () => inviteTableSubscriptionMock)
 mockClient.setRequestHandler(
   tablesQuery,
   tablesQueryMock.mockResolvedValue({
